+++ index.d.ts
@@ -1,0 +1,9 @@
+function find(type: "name" | "pid" | "port", value: string | number | RegExp, strict?: boolean): Promise<{
+    pid: number;
+    ppid?: number;
+    uid?: number;
+    gid?: number;
+    name: string;
+    cmd: string;
+}[]>
+export = find;